pyproject.toml
[project]
name = "mimicus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.1",
    "anyio>=4.3.0",
    "bcrypt>=4.1.2",
    "black>=24.2.0",
    "boto3>=1.34.50",
    "faker>=24.0.0",
    "fakeredis>=2.21.0",
    "fastapi>=0.110.0",
    "fastmcp>=2.14.2",
    "httpx>=0.27.0",
    "jinja2>=3.1.3",
    "jsonpath-ng>=1.6.1",
    "jsonschema>=4.21.1",
    "loguru>=0.7.2",
    "mypy>=1.9.0",
    "openapi-schema-validator>=0.6.2",
    "opentelemetry-api>=1.23.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-instrumentation-httpx>=0.44b0",
    "opentelemetry-sdk>=1.23.0",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=10.0.0",
    "prometheus-client>=0.20.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.1",
    "redis>=5.0.3",
    "regex>=2023.12.25",
    "restrictedpython>=7.0",
    "rich>=13.7.1",
    "ruff>=0.3.4",
    "setuptools>=69.2.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.27",
    "sqlmodel>=0.0.16",
    "starlette>=0.36.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "typer>=0.12.0",
    "urllib3>=2.2.0",
    "uvicorn[standard]>=0.27.0",
    "watchfiles>=0.21.0",
    "wheel>=0.43.0",
]
